update sf instances
