+++ src/lib/orders/OrderDisplay.tsx
@@ -3,12 +3,13 @@ import type { HydratedOrder } from "./types";
 import { GPUS_PER_NODE } from "../constants";
 import dayjs from "dayjs";
 import { formatDuration } from ".";
+import { Row } from "../Row";
+
 
 function Order(props: { order: HydratedOrder }) {
     const duration = dayjs(props.order.end_at).diff(props.order.start_at);
     const durationInHours = duration === 0 ? 1 : duration / 1000 / 60 / 60;
-    const pricePerGPUHour = props.order.price * props.order.quantity * GPUS_PER_NODE / durationInHours / 100;
-
+    const pricePerGPUHour = props.order.price * props.order.quantity / GPUS_PER_NODE / durationInHours / 100;
     const durationFormatted = formatDuration(duration);
 
     return (
@@ -19,18 +20,12 @@ function Order(props: { order: HydratedOrder }) {
                 <Text color={"yellow"} >{props.order.id}</Text>
                 <Text dimColor>({props.order.status})</Text>
             </Box>
-            <Box gap={1}>
-                <Text>{props.order.quantity * GPUS_PER_NODE}</Text>
-                <Text>{props.order.instance_type}</Text>
-                <Text>${pricePerGPUHour.toFixed(2)}/gpu/hr</Text>
-                <Text>{durationFormatted}</Text>
-                <Text>${props.order.price / 100}/total</Text>
-            </Box>
-            <Box gap={1}>
-                <Text>{dayjs(props.order.start_at).format("MMM D h:mm a").toLowerCase()}</Text>
-                <Text>{"→"}</Text>
-                <Text>{dayjs(props.order.end_at).format("MMM D h:mm a").toLowerCase()}</Text>
-            </Box>
+            <Row headWidth={7} head="nodes" value={`${props.order.quantity} x ${props.order.instance_type} (${props.order.quantity * GPUS_PER_NODE} gpus)`} />
+            <Row headWidth={7} head="price" value={`$${pricePerGPUHour.toFixed(2)}/gpu/hr`} />
+            <Row headWidth={7} head="time" value={durationFormatted} />
+            <Row headWidth={7} head="total" value={`$${props.order.price / 100}/total`} />
+            <Row headWidth={7} head="start" value={dayjs(props.order.start_at).format("MMM D h:mm a").toLowerCase()} />
+            <Row headWidth={7} head="end" value={dayjs(props.order.end_at).format("MMM D h:mm a").toLowerCase()} />
         </Box>
     );
 }
